Поменял тип на вещественный, поменял отладочный вывод

+++ lab_2/lab_2/point.h
@@ -1,17 +1,16 @@
 #pragma once
-#include <cstdint>
 #include <string>
 
 class Point {
 public:
   Point();
-  Point(int64_t x, int64_t y);
-  int64_t GetX() const { return x_; }
-  int64_t GetY() const { return y_; }
+  Point(double x, double y);
+  double GetX() const { return x_; }
+  double GetY() const { return y_; }
   std::string ToString() const;
   ~Point();
 
 private:
-  int64_t x_ = 0;
-  int64_t y_ = 0;
+  double x_ = 0;
+  double y_ = 0;
 };
